# refactor: use pcre regexp engine for custom rule